+++ menu.d.ts
@@ -1,0 +1,6 @@
+declare global {
+  interface HTMLElementTagNameMap {
+    "m-menu": HTMLElement & { refresh(): void };
+  }
+}
+export {};